Factor duplicated private methods into a static class
